Resolve #2407: align JWT refresh token provider surface
