+++ packages/jwt/src/module.test.ts
@@ -7,26 +7,10 @@ import { FluoFactory } from '@fluojs/runtime';
 
 import * as jwtRootExports from './index.js';
 import { JwtModule } from './module.js';
-import { type RefreshTokenRecord, type RefreshTokenStore, RefreshTokenService } from './refresh/refresh-token.js';
+import { RefreshTokenService } from './refresh/refresh-token.js';
 import { JwtService } from './service.js';
 import { DefaultJwtSigner } from './signing/signer.js';
 import { DefaultJwtVerifier } from './signing/verifier.js';
-
-class NoopRefreshTokenStore implements RefreshTokenStore {
-  async save(_: RefreshTokenRecord): Promise<void> {}
-
-  async find(_: string): Promise<RefreshTokenRecord | undefined> {
-    return undefined;
-  }
-
-  async revoke(_: string): Promise<void> {}
-
-  async revokeBySubject(_: string): Promise<void> {}
-
-  async consume(): Promise<'consumed' | 'already_used' | 'expired' | 'not_found' | 'mismatch'> {
-    return 'not_found';
-  }
-}
 
 @Inject(DefaultJwtSigner, DefaultJwtVerifier)
 class JwtRoundTripService {
@@ -253,28 +237,6 @@ describe('JwtModule', () => {
     }
   });
 
-  it('registers refresh token service for async options when refreshToken is configured', async () => {
-    const app = await createJwtApplicationContext(JwtModule.forRootAsync({
-      useFactory: async () => ({
-        algorithms: ['HS256'],
-        refreshToken: {
-          expiresInSeconds: 60,
-          rotation: true,
-          secret: 'refresh-secret',
-          store: new NoopRefreshTokenStore(),
-        },
-        secret: 'jwt-secret',
-      }),
-    }));
-
-    try {
-      expect(app.container.has(RefreshTokenService)).toBe(true);
-      await expect(app.container.resolve(RefreshTokenService)).resolves.toBeInstanceOf(RefreshTokenService);
-    } finally {
-      await app.close();
-    }
-  });
-
   it('exports refresh token service from async registration metadata to match sync registration parity', () => {
     const moduleType = JwtModule.forRootAsync({
       useFactory: async () => ({
@@ -288,52 +250,30 @@ describe('JwtModule', () => {
     expect(moduleProviders(moduleType).map((provider) => providerToken(provider))).toContain(RefreshTokenService);
   });
 
-  it('fails async bootstrap when refreshToken is configured without any HMAC algorithms', async () => {
-    await expect(createJwtApplicationContext(JwtModule.forRootAsync({
-      useFactory: async () => ({
-        algorithms: ['RS256'],
-        publicKey: '-----BEGIN PUBLIC KEY-----\nMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEApseudo\n-----END PUBLIC KEY-----',
-        refreshToken: {
-          expiresInSeconds: 60,
-          rotation: false,
-          secret: 'refresh-secret',
-          store: new NoopRefreshTokenStore(),
-        },
-      }),
-    }))).rejects.toThrow(
-      'JWT refresh token verifier requires at least one HMAC algorithm (HS256/HS384/HS512) in the allowed algorithms list.',
-    );
-  });
-
-  it('registers refresh token service when refresh options are provided', async () => {
-    const container = new Container();
+  it('exports refresh token service from synchronous registration metadata when refresh options are omitted', () => {
     const moduleType = JwtModule.forRoot({
       algorithms: ['HS256'],
-      refreshToken: {
-        expiresInSeconds: 60,
-        rotation: true,
-        secret: 'refresh-secret',
-        store: new NoopRefreshTokenStore(),
-      },
-      secret: 'jwt-secret',
-    });
-
-    container.register(...moduleProviders(moduleType));
-    await expect(container.resolve(RefreshTokenService)).resolves.toBeInstanceOf(RefreshTokenService);
-  });
-
-  it('exports refresh token service from synchronous registration when refresh options are provided', () => {
-    const moduleType = JwtModule.forRoot({
-      algorithms: ['HS256'],
-      refreshToken: {
-        expiresInSeconds: 60,
-        rotation: true,
-        secret: 'refresh-secret',
-        store: new NoopRefreshTokenStore(),
-      },
-      secret: 'jwt-secret',
+      issuer: 'jwt-module-tests',
+      secret: 'sync-secret-without-refresh',
     });
 
     expect(moduleExports(moduleType)).toContain(RefreshTokenService);
+    expect(moduleProviders(moduleType).map((provider) => providerToken(provider))).toContain(RefreshTokenService);
+  });
+
+  it('rejects sync refresh token service resolution when sync options omit refreshToken', async () => {
+    const app = await createJwtApplicationContext(JwtModule.forRoot({
+      algorithms: ['HS256'],
+      issuer: 'jwt-module-tests',
+      secret: 'sync-secret-without-refresh',
+    }));
+
+    try {
+      await expect(app.container.resolve(RefreshTokenService)).rejects.toThrow(
+        'JWT refresh token options are not configured.',
+      );
+    } finally {
+      await app.close();
+    }
   });
 });
